Move backend specific utility function to backends.utils
