+++ silx/gui/plot/backends/BackendMatplotlib.py
@@ -52,7 +52,7 @@ from matplotlib.backend_bases import MouseEvent
 from matplotlib.lines import Line2D
 from matplotlib.collections import PathCollection, LineCollection
 
-from .. import _utils
+from . import utils
 from .ModestImage import ModestImage
 from . import BackendBase
 from .. import Colors
@@ -543,7 +543,7 @@ class BackendMatplotlib(BackendBase.BackendBase):
                 ymin2, ymax2 = ranges.yright
 
             # Add margins around data inside the plot area
-            newLimits = list(_utils.addMarginsToLimits(
+            newLimits = list(utils.addMarginsToLimits(
                 dataMargins,
                 self.ax.get_xscale() == 'log',
                 self.ax.get_yscale() == 'log',
